+++ src/app/layout.tsx
@@ -3,6 +3,7 @@ import { Geist, Geist_Mono } from 'next/font/google';
 import './globals.css';
 import { Providers } from './providers/providers';
 import AuthHeaderWrapper from '../components/layout/AuthHeaderWrapper';
+import { getCountries } from '../actions/countries';
 
 const geistSans = Geist({
   variable: '--font-geist-sans',
@@ -19,15 +20,16 @@ export const metadata: Metadata = {
   description: 'Generated by create next app',
 };
 
-export default function RootLayout({
+export default async function RootLayout({
   children,
 }: Readonly<{
   children: React.ReactNode;
 }>) {
+  const countries = await getCountries();
   return (
     <html lang="en" suppressHydrationWarning>
       <body className={`${geistSans.variable} ${geistMono.variable} antialiased`}>
-        <Providers>
+        <Providers countries={countries}>
           <AuthHeaderWrapper>
             <div className="p-8">{children}</div>
           </AuthHeaderWrapper>
